Break EntityFormTest into smaller test files
